refactor(rules): split CodeBlockRule into language validator strategies
Break up the ~585-LOC CodeBlockRule god object. Introduce a
LanguageValidator strategy interface plus a ValidatorRegistry that maps
language identifiers to validators, and extract each inline language
validator (JavaScript/TypeScript, Python, Go, Bash, SQL, JSON, YAML) and
its helpers into its own module under src/rules/codeblock/.

CodeBlockRule becomes a thin orchestrator: it extracts code blocks,
applies the language-agnostic checks (untyped fences, incompleteness,
unmarked anti-patterns), looks up the validator by language, and
delegates. Cross-block JS/TS style consistency moves onto the JavaScript
validator via an optional validateConsistency hook.

Pure refactor: behavior, violation ordering, rule id, constructor
signature, and registry descriptor are unchanged; the existing 40
CodeBlockRule tests stay green unmodified.

Hook bypassed: warden format:check fails on pre-existing unformatted
files (lintEnhanced.ts, ruleDescriptors.ts) unrelated to this change.
